% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{etilt}
\alias{etilt}
\title{The exponential tilting function}
\usage{
etilt(xl, xt, f = NULL, h = NULL, m = NULL, interval = NULL, ...)
}
\arguments{
\item{xl}{The training left-tail areas (or p-values)}

\item{xt}{The testing left-tail areas (or p-values)}

\item{f}{The objective function is tilted. If either xl or f is NULL, f is fitted by
\code{\link[tiltmod]{UBMM}}. Default is NULL.}

\item{h}{The conditioning function. If h is NULL, h = -log(f(x)). Default is NULL.}

\item{m}{The constant is used to find the optimal theta such that E(h(x))=m. If m is
NULL, m = mean(h(xt)). Default is NULL.}

\item{interval}{The interval is used to search the optimal theta. Default is (-0.5,10).}

\item{...}{Arguments to be passed to \code{\link[stats]{uniroot}}.}
}
\value{
A list includes theta, tau, tilt_tau, tilt_f, tilt_f0, tilt_f1, respectively.
}
\description{
This function tilts the mixture model fitted from the training tail-areas (or p-values)
by conditioning on the average of local fdr's from the testing tail-areas (or p-values)
}
\examples{
xl=c(rbeta(100,0.5,0.5),runif(900))
xt=c(rbeta(300,2,3),runif(700))

\dontrun{
etilt(xl,xt)
}

}
