% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power}
\alias{power}
\title{Empirical power analysis}
\usage{
power(q, x, w = NULL, a = NULL, precision = 1e-08, MaxIter = 10000L,
  theta = NULL, alpha = 0.9, type = c("left tail area", "pvalue"),
  rel.tol = .Machine$double.eps^0.25, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{q}{A numeric value or a vector of numerical value, represent global false 
discovery rate used for power analysis.}

\item{x}{A vector of numeric values, represent p-values or left-tail areas 
of test statistics from a differential gene expression study.}

\item{w}{A vector of two numeric values, represent the weights
of the uniform and Beta distributions. See \code{\link[tiltmod]{UBMM}}.}

\item{a}{A vector of two initial parameter values for Beta distribution. See \code{\link[tiltmod]{UBMM}}.}

\item{precision}{The precision for convergence. Default value is 1e-8.}

\item{MaxIter}{The maximum iteration for the EM algorhthm.}

\item{theta}{A numerical value, represents the exponential tilting parameter 
for the fitted mixture model from x. Defualt is NULL.}

\item{alpha}{A numeric value, used to determine the probably null region in method ``m1" 
(see \code{\link[tiltmod]{tqvalue}}). Default is 0.9.}

\item{type}{A character value, chosen from ``left tail area'' and ``pvalue''. Default is
``left tail area''.}

\item{rel.tol}{the accuracy used in \code{\link[stats]{integrate}}.}

\item{tol}{the accuracy used in \code{\link[stats]{uniroot}}.}
}
\value{
A dataframe consists of q, TypeI, Power, ProbS, respectively. TypeI, Power,
        and ProbS are calculated based on the rejection region R(q) and
        the empirical mixture model for x. 
        \describe{
           \item{q}{The global false discovery rates provided in arguments.}
           \item{TypeI}{\eqn{P(R(q)|H_0)}}
           \item{power}{\eqn{P(R(q)|H_1)}}
           \item{ProbS}{\eqn{P(R(q))}}
        }
        
        If theta is provided, then the results contain two data frames as above, one 
        is calculated from the non-exponential tilted mixture model and the other 
        from the exponential tilted mixture model, respectively.
}
\description{
This function deals with power analysis by calculating the relevant TypeI error,
power, and probability of being significant by given global false discovery rate.
}
\examples{
x=c(rbeta(50,0.5,0.5),runif(950))
q=seq(0.05,0.95,0.05)
\dontrun{
power(q,x,alpha=0.9,type="left tail area")
power(q,x,theta=2,alpha=0.9,type="left tail area")
}

}
