% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tqvalue}
\alias{tqvalue}
\title{The exponential tilting mixture model}
\usage{
tqvalue(xl, xt, precision = 1e-06, MaxIter = 10000, ncores = 1,
  adjust = FALSE, method = c("m1", "m2", "m3"), interval = NULL,
  type = c("left tail area", "pvalue"), c = 1, alpha = 0.1,
  rel.tol = .Machine$double.eps^0.5, tol = 1e-05)
}
\arguments{
\item{xl}{The training left-tail areas (or p-values)}

\item{xt}{The testing left-tail areas (or p-values)}

\item{precision}{The precision for convergence. Default value is 1e-6.}

\item{MaxIter}{The maximum iteration for the EM algorhthm.}

\item{ncores}{The number of cpus used for implementing this function.}

\item{adjust}{Whether or not to do the model adjustment. Default is false.}

\item{method}{A character chosen from m1, m2, m3. Default is m1.}

\item{interval}{A vector of two numeric values, which determines the range to
search the optimal theta. Default is c(-100L,100L).}

\item{type}{A character value, chosen from left tail area and pvalue. Default is
left tail area.}

\item{c}{A numeric value. Used in method "m1", a threshold to segment the signal and
baseline in f1. Default is 1.}

\item{alpha}{A numeric value. Used in method "m2" and "m3", an arbitrary value to
determine the contamination area of signal and baseline in f.}

\item{rel.tol}{the accuracy used in \code{\link[stats]{integrate}}.}

\item{tol}{the accuracy used in \code{\link[stats]{uniroot}}.}
}
\value{
A dataframe includes xl, xt, fdr, FDR, tfdr, and tFDR, respectively. fdr and
        FDR are the local and global false discovery rate for each value of xt.
        tfdr and tFDR are the corresponding tilted local and global false discovery
        rate, respectively.

        The optimal theta calculated by solving log(E(exp(thetah(x))))-ctheta,
        where c=mean(h(xt)).
}
\description{
This function tilts the mixture model fitted from the training tail-areas (or p-values)
by conditioning on the average of local fdr's from the testing tail-areas (or p-values)
}
\examples{
xl=c(rbeta(50,0.2,0.2),runif(950))
xt=c(rbeta(50,0.1,0.1),runif(950))

\dontrun{
tqvalue(xl,xt,ncores=4,adjust=FALSE,type="left tail area")
}

}
