% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tqvalue2}
\alias{tqvalue2}
\title{The exponential tilting adjusted mixture model}
\usage{
tqvalue2(xl, xt, precision = 1e-06, MaxIter = 10000, ncores = 1,
  method = c("m1", "m2", "m3"), interval = NULL,
  type = c("left tail area", "pvalue"), c = 1, alpha = 0.1,
  rel.tol = .Machine$double.eps^0.5)
}
\arguments{
\item{xl}{The training left tail areas (or p-values)}

\item{xt}{The testing left tail areas (or p-values)}

\item{precision}{The precision for convergence. Default value is 1e-6.}

\item{MaxIter}{The maximum iteration for the EM algorhthm.}

\item{ncores}{The number of cpus used for implementing this function.}

\item{method}{A character chosen from m1, m2, m3. Default is m1.}

\item{interval}{A vector of two numeric values, which determines the range to
search the optimal theta. Default is c(-100L,100L).}

\item{type}{A character value, chosen from left tail area and pvalue. Default is
left tail area.}

\item{c}{A numeric value. Used in method "m1", a threshold to segment the signal and
baseline in f1. Default is 1.}

\item{alpha}{A numeric value. Used in method "m2" and "m3", an arbitrary value to
determine the contamination area of signal and baseline in f.}
}
\value{
A dataframe includes xl, xt, FDR, and ttFDR, respectively. FDR is the
        global false discovery rate for each value of xt, and ttFDR is the tilted
        global fdr.

        The optimal theta calculated by solving log(E(exp(thetah(x))))-ctheta,
        where c=mean(h(xt)).
}
\description{
This function tilts the mixture model fitted from the training tail-areas (or p-values)
by conditioning on the average of local fdr's from the testing tail-areas (or p-values)
}
\examples{
xl=c(rbeta(50,0.2,0.2),runif(950))
xt=c(rbeta(50,0.1,0.1),runif(950))

\dontrun{
tqvalue2(xl,xt,ncores=4)
}

}
