% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UBMM}
\alias{UBMM}
\title{Fit a mixture of uniform and Beta distribution}
\usage{
UBMM(x, w = as.numeric(c()), a = as.numeric(c()), precision = 1e-08,
  MaxIter = 10000L)
}
\arguments{
\item{x}{A vector of numeric values}

\item{w}{A vector of two numeric values, representing the weights
of the uniform and Beta distributions. Default 
values are 0.5, respectively.}

\item{a}{Initial values of the alpha and beta for the Beta
distribution. Defaults are obtained from MOM estimators.}

\item{precision}{The tolerance for convergence. Default value is 
1e-8.}

\item{MaxIter}{The maximum iteration for the EM algorithm. Default value
is 10000L.}
}
\value{
A list of three components, including the converged weight,
        parameters for Beta distribution, and the convergence iteration, 
        respectively.
}
\description{
Fit a mixture of uniform and Beta distribution
}
\examples{
x0=runif(900) 
x1=rbeta(100,0.5,0.5)
UBMM(c(x0,x1),w=c(0.8,0.2),a=c(0.7,0.8))

}
