% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fnet}
\alias{fnet}
\title{Frequency network}
\usage{
fnet(x, Simplify = FALSE, threshold = 0.05, max.ew = 3, max.size = 24,
  node = TRUE, directed = FALSE, print.graph = TRUE, ...)
}
\arguments{
\item{x}{A list of discovered differentially expressed genes}

\item{Simplify}{Logical indicating whether to discard the genes with lower relative
freqency than the threshold. Default is FALSE.}

\item{threshold}{A numeric value determining the cutoff point, where the genes
are discarded with lower relative frequency than it. Default is 0.05.}

\item{max.ew}{A numeric value. The maximum edge width in the network plot.}

\item{max.size}{A numeric value, displays the maximum node size.}

\item{node}{A logical value, illustrates the feature name in a circle. Default is TRUE.}

\item{directed}{A logical value, indicates whether the edges are shown in directions. Default 
is FALSE.}

\item{print.graph}{A logical value. Default is TRUE. If FALSE, then do not print 
the frequency network.}

\item{...}{Arguments to be passed to \code{\link[igraph]{plot.igraph}}.}
}
\value{
A network plot.
}
\description{
This function displays the frequency network of discovered differentially expressed
genes.
}
\examples{
x=list(c(1,3,4),c(2,4,5),c(3,5,1),c(4,1,2),c(5,2,3))
\dontrun{
require(igraph)
fnet(x)
fnet(x,layout=layout.fruchterman.reingold,vertex.color="grey60")
}

}
