% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Marfan}
\alias{Marfan}
\title{Marfan}
\format{A data frame for 101 samples with 4133 variables: \code{treatment}, 
  \code{X1},...,\code{X4132}. \code{treatment} contains 41 samples from 
  the control group and 60 samples from the Marfan group.}
\usage{
Marfan
}
\description{
Gene expression data includes the treatment group and 4132 gene expressions.
}
\keyword{datasets}
